+++ core/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: 362b999b90d53859aa7b926a59c970f3ea31abf4
+  channel: dev
+
+project_type: app
