+++ core/README.md
@@ -1,0 +1,9 @@
+# shared
+
+A new Flutter project.
+
+## Getting Started
+
+To update the auto generated json run  `flutter packages pub run build_runner watch  —delete-conflicting-outputs`
+
+To updated the models for the website push the project to GitHub and run packages get on the website. You can update the override [here](/website/pubspec.yaml)
