+++ shared/lib/shared.dart
@@ -1,19 +1,4 @@
 library shared;
 
 export 'src/index.dart';
-
-class Shared {
-  Shared._();
-
-  static bool testMethod() {
-    return true;
-  }
-
-  static int incrementValue(int value) {
-    return value + 1;
-  }
-
-  static int decrementValue(int value) {
-    return value - 1;
-  }
-}
+export 'package:shared_core/shared.dart';
