+++ shared/lib/src/index.dart
@@ -1,6 +1,4 @@
 export 'utils/index.dart';
 export 'models/index.dart';
-export 'classes/index.dart';
-export 'abstract/index.dart';
 export 'locale/index.dart';
 export 'constants/index.dart';
