+++ shared/test/widget_test.dart
@@ -1,0 +1,18 @@
+// This is a basic Flutter widget test.
+//
+// To perform an interaction with a widget in your test, use the WidgetTester
+// utility that Flutter provides. For example, you can send tap and scroll
+// gestures. You can also use WidgetTester to find child widgets in the widget
+// tree, read text, and verify that the values of widget properties are correct.
+
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:shared/shared.dart';
+
+void main() {
+  test('Counter increments smoke test', (_) {
+    // Verify that our counter starts at 0.
+    expect(Shared.incrementValue(1), 2);
+    expect(Shared.decrementValue(1), 0);
+  });
+}
